+++ frontend/src/components/SignUpForm.jsx
@@ -347,6 +347,7 @@ function SignUpComponent() {
                         {isSubmitting ? '가입 처리 중..' : '회원가입'}
                     </button>
                   </form>
+                  <div style={{ height: "30px" }}></div>
                 </>
             );
         }
