+++ frontend/src/layouts/find.css
@@ -1,8 +1,9 @@
 .auth-page {
-  min-height: calc(100vh - 65px); 
+  min-height: calc(100vh - 70px); 
   display: flex;
   justify-content: center;
-  align-items: center;
+  /* align-items: center; */
+  /* margin-top: 180px; */
   background-color: #000; 
   color: #f5f5f5;
   padding: 40px 16px;
@@ -11,9 +12,10 @@
 .auth-card {
   width: 100%;
   max-width: 420px;
+  margin-top: 180px;
   /* background: #151515; */
   border-radius: 14px;
-  padding: 28px 26px 26px;
+  padding: 26px 26px;
   box-shadow: 0 10px 30px rgba(0, 0, 0, 0.7);
   /* border: 1px solid #2a2a2a; */
 }
@@ -27,7 +29,7 @@
 .auth-subtitle {
   margin-top: 10px;
   margin-bottom: 20px;
-  font-size: 13px;
+  font-size: 15px;
   color: #aaaaaa;
   line-height: 1.4;
 }
@@ -43,14 +45,14 @@
 }
 
 .auth-field label {
-  font-size: 12px;
+  font-size: 15px;
   color: #888;
   margin-bottom: 6px;
 }
 
 .auth-input {
-  padding: 10px 12px;
-  font-size: 14px;
+  padding: 12px 12px;
+  font-size: 15px;
   border-radius: 6px;
   border: 1px solid #444;
   background-color: #111;
@@ -75,10 +77,10 @@
 
 .auth-btn {
   width: 100%;
-  height: 40px;
+  height: 45px;
   border-radius: 8px;
   border: none;
-  font-size: 14px;
+  font-size: 15px;
   cursor: pointer;
   margin-top: 4px;
 }
@@ -86,7 +88,6 @@
 .auth-btn.primary {
   background: #f4002d;
   color: #000;
-  font-weight: 600;
 }
 
 .auth-btn.primary:hover {
@@ -118,7 +119,7 @@
   display: flex;
   align-items: center;
   gap: 6px;
-  font-size: 12px;
+  font-size: 15px;
   color: #777;
 }
 
@@ -134,7 +135,7 @@
   display: inline-flex;
   align-items: center;
   justify-content: center;
-  font-size: 11px;
+  font-size: 15px;
 }
 
 .step-item.active .step-circle {
@@ -163,5 +164,5 @@
 .bottom-links .link {
   color: #f5f5f5;
   text-decoration: none;
-  font-size: 14px;
+  font-size: 15px;
 }
