+++ frontend/src/pages/AdminLogin.jsx
@@ -1,10 +1,8 @@
-import Header from "../layouts/usermain/header";
 import AdminLoginBody from "../layouts/AdminLoginMain.jsx";
 
 function AdminLoginMain (){
     return(
         <>
-        <Header/>
         <AdminLoginBody/>
         </>
     )
